First commit for transferCacheToSDCard.sh
This script presents a menu to select the game to transfer to MicroSD card. It reads the app manifest files to tell which games are installed in the MicroSD card and builds the menu from that data.

The menu isn't the prettiest, but it works. Bash arrays can be a bit of a pain, so I am not sure if I will fix the way the games are listed in the menu.

Anyway, this isn't fully tested yet, so I don't recommend using it yet. Committing it to have a safe copy over on Github and for visitors to know that I'm working on this new script!